\name{biglasso-internal}

\title{Internal biglasso functions}
\alias{loss}

\description{Internal biglasso functions}

\usage{
loss.biglasso(y, yhat, family, eval.metric)
}

\author{
  Yaohui Zeng and Patrick Breheny
  
  Maintainer: Yaohui Zeng <yaohui-zeng@uiowa.edu>
}

\details{
These are not intended for use by users.\code{loss.biglasso} 
calculates the value of the loss function for the given predictions (used for
cross-validation).
}

\keyword{internal}
