% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.cox.R
\name{deviance.cox}
\alias{deviance.cox}
\title{Internal biglasso functions}
\usage{
\method{deviance}{cox}(X, y, beta)
}
\arguments{
\item{X}{The design matrix, without an intercept. It must be a
double type \code{\link[bigmemory]{big.matrix}} object.}

\item{y}{A response vector.}

\item{beta}{Fitted coefficients obtained from a \code{\link[biglasso]{biglasso}} object.}
}
\value{
todo...
}
\description{
Internal biglasso functions
}
\details{
These are not intended for use by users.\code{loss.biglasso} calculates the 
value of the loss function for the given predictions (used for cross-validation).
}
\author{
David Fleischer

Maintainer: Yaohui Zeng <yaohui.zeng@gmail.com>
}
