% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss.biglasso}
\alias{loss.biglasso}
\title{Internal biglasso functions}
\usage{
loss.biglasso(y, yhat, family, eval.metric, grouped = TRUE)
}
\arguments{
\item{y}{The observed response vector.}

\item{yhat}{The predicted response vector.}

\item{family}{Either "gaussian" or "binomial", depending on the response.}

\item{eval.metric}{The evaluation metric for the cross-validated error and
for choosing optimal \code{lambda}:
\itemize{
\item \code{"default"} for linear regression is MSE (mean squared error) and for logistic regression is
misclassification error.
\item \code{"MAPE"}, for linear regression only, is the Mean Absolute Percentage Error.
\item "auc", for logistic regression, is the area under the receiver operating
characteristic curve (ROC).
}}

\item{grouped}{Whether to calculate loss for the entire CV fold (\code{TRUE}), or for predictions
individually. Must be \code{TRUE} when \code{eval.metric} is 'auc'.}
}
\description{
Internal biglasso functions
}
\details{
These are not intended for use by users. \code{loss.biglasso} calculates the value of the loss
function for the given predictions (used for cross-validation).
}
\keyword{internal}
