% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mbiglasso.R
\name{plot.mbiglasso}
\alias{plot.mbiglasso}
\title{Plot coefficients from a "mbiglasso" object}
\usage{
\method{plot}{mbiglasso}(x, alpha = 1, log.l = TRUE, norm.beta = TRUE, ...)
}
\arguments{
\item{x}{Fitted \code{mbiglasso} model.}

\item{alpha}{Controls alpha-blending, helpful when the number of covariates is large. Default is
alpha=1.}

\item{log.l}{Should horizontal axis be on the log scale? Default is TRUE.}

\item{norm.beta}{Should the vertical axis be the l2 norm of coefficients for each variable?
Default is TRUE. If False, the vertical axis is the coefficients.}

\item{...}{Other graphical parameters to \code{\link[=plot]{plot()}}}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of producing a plot.
}
\description{
Produce a plot of the coefficient paths for a fitted multiple responses \code{mbiglasso} object.
}
\seealso{
\code{\link[=biglasso]{biglasso()}}
}
