% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.deviance.R
\name{cox.deviance}
\alias{cox.deviance}
\title{deviance and log-likelihood calculations for fitted Cox models}
\usage{
cox.deviance(X, y, beta, row.idx)
}
\arguments{
\item{X}{The design matrix, without an intercept. It must be a
double type \code{\link[bigmemory]{big.matrix}} object.}

\item{y}{A response vector.}

\item{beta}{Fitted coefficients obtained from a \code{\link[biglasso]{biglasso}} object.}

\item{row.idx}{Integer vector of row/observation indices to be used in computing model likelihoods.}
}
\value{
A list with deviance and log-likelihood values calculated over the fitted coefficients.
\item{nulldev}{The null deviance of the model defined to be -2 * (loglike_sat - loglike_null), where the null
model in the context of Cox regression refers to the 0 model.} \item{dev}{Fitted deviance associated with
the observations and fitted coefficients. Defined to be -2 * (loglike_sat - loglike), where loglike
is calculated using the Breslow approximation of the Cox model likelihood.} \item{dev.ratio}{The fraction of
(null) deviance explained, dev.ratio = 1 - dev/nulldev. Calculated primarily as a way to quickly compare
outputs with those of \code{"glmnet"}.}
}
\description{
Calculate the deviances (model, null) and log-likelihoods (saturated, model, null) 
associated with a set of observations and fitted coefficients.
}
\author{
David Fleischer

Maintainer: Yaohui Zeng <yaohui.zeng@gmail.com>
}
