% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss.biglasso}
\alias{loss.biglasso}
\title{Internal biglasso functions}
\usage{
loss.biglasso(y, yhat, family, eval.metric)
}
\arguments{
\item{y}{The observed response vector.}

\item{yhat}{The predicted response vector.}

\item{family}{Either "gaussian" or "binomial", depending on the response.}

\item{eval.metric}{The evaluation metric for the cross-validated error and
for choosing optimal \code{lambda}. "default" for linear regression is MSE
(mean squared error), for logistic regression is misclassification error.
"MAPE", for linear regression only, is the Mean Absolute Percentage Error.}
}
\description{
Internal biglasso functions
}
\details{
These are not intended for use by users.\code{loss.biglasso} calculates the 
value of the loss function for the given predictions (used for cross-validation).
}
\author{
Yaohui Zeng and Patrick Breheny

Maintainer: Yaohui Zeng <yaohui.zeng@gmail.com>
}
\keyword{internal}
