% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.biglasso.R
\name{plot.biglasso}
\alias{plot.biglasso}
\title{Plot coefficients from a "biglasso" object}
\usage{
\method{plot}{biglasso}(x, alpha = 1, log.l = TRUE, ...)
}
\arguments{
\item{x}{Fitted \code{"biglasso"} model.}

\item{alpha}{Controls alpha-blending, helpful when the number of covariates
is large.  Default is alpha=1.}

\item{log.l}{Should horizontal axis be on the log scale?  Default is TRUE.}

\item{\dots}{Other graphical parameters to \code{plot}}
}
\description{
Produce a plot of the coefficient paths for a fitted \code{\link{biglasso}}
object.
}
\examples{

## See examples in "biglasso"

}
\seealso{
\code{\link{biglasso}}, \code{\link{cv.biglasso}}
}
\author{
Yaohui Zeng and Patrick Breheny

Maintainer: Yaohui Zeng <yaohui.zeng@gmail.com>
}
\keyword{models}
\keyword{regression}
